requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.13
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-ipware==2.1.0
django-js-asset==1.2.2
django-recaptcha3==0.2.0
django-tenant-schemas==1.10.0
djangorestframework==3.11.0
easy-thumbnails==2.7
idna==2.10
jsonfield==3.0.0
nepalicalendar==0.4
ordered-set==4.0.2
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
